requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
psycopg2-binary==2.9.6
pydantic==2.0.0